+++ packages/checkbox-group/src/Component.responsive.tsx
@@ -1,6 +1,7 @@
 import React, { FC } from 'react';
 
 import { useMatchMedia } from '@alfalab/core-components-mq';
+import { getComponentBreakpoint } from '@alfalab/core-components-shared';
 
 import { BaseCheckboxGroupProps } from './components/base-checkbox-group';
 import { CheckboxGroupDesktop } from './desktop';
@@ -20,7 +21,7 @@ export type CheckboxGroupProps = Omit<BaseCheckboxGroupProps, 'styles'> & {
 };
 
 export const CheckboxGroup: FC<CheckboxGroupProps> = ({
-    breakpoint = 1024,
+    breakpoint = getComponentBreakpoint(),
     defaultMatchMediaValue,
     ...restProps
 }) => {