+++ packages/gallery/src/components/header-mobile/index.ts
@@ -1,0 +1,1 @@
+export { HeaderMobile } from './Component';