+++ packages/gallery/src/components/index.ts
@@ -2,3 +2,5 @@ export * from './navigation-bar';
 export * from './header';
 export * from './image-preview';
 export * from './image-viewer';
+export * from './info-bar';
+export * from './header-mobile';