+++ packages/gallery/src/utils/utils.ts
@@ -10,9 +10,11 @@ export const getImageAlt = ({ alt, name }: GalleryImage, index: number): string 
     alt || name || `Изображение ${index + 1}`;
 
 export const isSmallImage = (meta?: ImageMeta) => {
-    if (!meta) {
-        return false;
+    if (meta && meta.width && meta.height) {
+        return meta.width < PLACEHOLDER_WIDTH && meta.height < PLACEHOLDER_HEIGHT;
     }
 
-    return meta.width < PLACEHOLDER_WIDTH && meta.height < PLACEHOLDER_HEIGHT;
+    return false;
 };
+
+export const isVideo = (url: string | undefined) => /(.*?)(\.webm|\.mp4|\.m3u8)$/.test(url ?? '');