+++ packages/modal/src/Component.responsive.tsx
@@ -1,6 +1,7 @@
 import React, { forwardRef } from 'react';
 
 import { useMatchMedia } from '@alfalab/core-components-mq';
+import { getComponentBreakpoint } from '@alfalab/core-components-shared';
 
 import { Content } from './components/content/Component';
 import { Controls } from './components/controls';
@@ -10,7 +11,15 @@ import { Modal } from './Component';
 import type { ModalResponsiveProps } from './typings';
 
 const ModalResponsiveComponent = forwardRef<HTMLDivElement, ModalResponsiveProps>(
-    ({ children, breakpoint = 1024, defaultMatchMediaValue = true, ...restProps }, ref) => {
+    (
+        {
+            children,
+            breakpoint = getComponentBreakpoint(),
+            defaultMatchMediaValue = true,
+            ...restProps
+        },
+        ref,
+    ) => {
         const [isDesktop] = useMatchMedia(`(min-width: ${breakpoint}px)`, defaultMatchMediaValue);
 
         return (