+++ packages/shared/src/index.ts
@@ -10,3 +10,4 @@ export * from './os';
 export * from './dom';
 export * from './exhaustiveCheck';
 export * from './context/PortalContext';
+export * from './getComponentBreakpoint';