+++ packages/amount/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 3.8.0
+
+### Minor Changes
+
+<sup><time>20.09.2024</time></sup>
+
+### [#1213](https://github.com/core-ds/core-components/pull/1213)
+
+-   Добавлена возможность глобального переопределения breakpoint компонентов [Документация](https://core-ds.github.io/core-components/master/?path=/docs/instructions-breakpoints--docs)
+
 ## 3.7.0
 
 ### Minor Changes