+++ packages/intl-phone-input/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log
 
+## 9.10.5
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   select@17.14.1
+    -   input-autocomplete@12.5.1
+
+## 9.10.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input-autocomplete@12.5.0
+    -   select@17.14.0
+
+## 9.10.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input-autocomplete@12.4.0
+    -   select@17.13.0
+
 ## 9.10.2
 
 ### Patch Changes