+++ packages/pagination/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 2.4.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   tag@8.5.0
+
 ## 2.4.1
 
 ### Patch Changes