+++ packages/pass-code-v1/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @alfalab/core-components-pass-code-v1
 
+## 1.1.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.9.0
+    -   shared@0.13.0
+
+## 1.1.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.8.0
+
 ## 1.1.2
 
 ### Patch Changes