+++ packages/textarea/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 8.8.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   form-control@12.6.0
+    -   shared@0.13.0
+
+## 8.8.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   form-control@12.5.0
+
 ## 8.8.1
 
 ### Patch Changes